requirements.txt
altair==5.0.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
ffmpeg-python==0.2.0
frozendict==2.3.8
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.25.2
opencv-python==4.8.1.78
packaging==23.1
pandas==2.0.3
parso==0.8.3
peewee==3.17.0
pexpect==4.8.0
Pillow==9.5.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.9.2
six==1.16.0
smmap==5.0.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.21.2
wcwidth==0.2.9
webencodings==0.5.1
yfinance==0.2.31
zipp==3.16.2
